refactor: pending courses controller refactored
