+++ backend/src/controllers/pendingCourses/getPendingCourses/utils/mappers.ts
@@ -1,8 +1,7 @@
 import { NOTHING } from 'config/constants';
 import { ITimePeriod } from 'interfaces/common/datetime';
 import { IPendingCourse, TPendingCourses } from 'interfaces/Ientities/Iusers';
-
-import { convertToTimePeriod } from '../typeConversion/datetime/datetimeTypeConversions';
+import { convertToTimePeriod } from 'utils/typeConversion/datetime/datetimeTypeConversions';
 
 const mapPendingCourse = (course: IPendingCourse): IPendingCourse => {
   const { date: applicationDate } = course;
