+++ backend/src/controllers/pendingCourses/index.ts
@@ -1,0 +1,4 @@
+export { default as approvePendingCourse } from './approvePendingCourse/approvePendingCourse';
+export { default as declinePendingCourse } from './declinePendingCourse/declinePendingCourse';
+export { default as getPendingCourses } from './getPendingCourses/getPendingCourses';
+export { default as updatePendingCourse } from './updatePendingCourse/updatePendingCourse';
